requirements/dev.txt
-r base.txt

django-debug-toolbar==1.8
django-extensions==1.7.9
pygraphviz==1.3.1

# Test
factory-boy==2.8.1
django-test-plus=1.0.17
